rename UpdateCustomizationFieldsHandler to UpdateProductCustomizationFieldsHandler
